+++ lib/voyager/services/node_connector.ex
@@ -2,9 +2,6 @@ defmodule Voyager.Services.NodeConnector do
   @moduledoc "Connects to remote nodes via Erlang distribution."
 
   alias Voyager.Services.Distribution
-  alias Voyager.Settings
-
-  require Logger
 
   @spec connect(String.t(), String.t(), keyword()) :: {:ok, atom()} | {:error, term()}
   def connect(node_name, cookie, opts \\ []) do
@@ -31,64 +28,6 @@ defmodule Voyager.Services.NodeConnector do
   def disconnect(node) do
     Node.disconnect(node)
     :ok
-  end
-
-  defp ensure_distributed(name_type) when name_type in [:longnames, :shortnames] do
-    cond do
-      not Node.alive?() ->
-        start_distribution(name_type)
-
-      matches_name_type?(name_type) and distribution_name_matches?() ->
-        :ok
-
-      true ->
-        case :net_kernel.stop() do
-          :ok -> start_distribution(name_type)
-          {:error, reason} -> {:error, {:net_kernel_stop, reason}}
-        end
-    end
-  end
-
-  defp ensure_distributed(_name_type) do
-    {:error, :invalid_name_type}
-  end
-
-  defp matches_name_type?(:longnames), do: :net_kernel.longnames() == true
-  defp matches_name_type?(:shortnames), do: :net_kernel.longnames() == false
-
-  defp start_distribution(name_type) do
-    node_name = distribution_name()
-
-    case :net_kernel.start(node_name, %{name_domain: name_type, hidden: true}) do
-      {:ok, _pid} ->
-        :ok
-
-      {:error, {:already_started, pid}} ->
-        Logger.warning(
-          "net_kernel.start/2 returned {:already_started, #{inspect(pid)}} " <>
-            "for #{inspect(node_name)} name_type=#{inspect(name_type)}"
-        )
-
-        :ok
-
-      {:error, reason} ->
-        {:error, {:net_kernel, reason}}
-    end
-  end
-
-  defp distribution_name_matches? do
-    distributed_node_name =
-      Node.self()
-      |> Atom.to_string()
-      |> String.split("@", parts: 2)
-      |> hd()
-
-    distributed_node_name == Atom.to_string(distribution_name())
-  end
-
-  defp distribution_name do
-    suffix = Settings.get(:distribution_suffix, "")
-    String.to_atom("voyager#{suffix}")
   end
 
   defp diagnose_failure(node_name) do
